added navbar and routing with reach router
